# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GenMedia Vault",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/translations": {
      "get": {
        "tags": [
          "translations"
        ],
        "summary": "Browse",
        "operationId": "browse_api_v1_translations_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 500,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Group"
            }
          },
          {
            "name": "untranslated",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Untranslated"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "translations"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_translations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/translations/lookup": {
      "post": {
        "tags": [
          "translations"
        ],
        "summary": "Lookup",
        "operationId": "lookup_api_v1_translations_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginInput": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginInput"
      },
      "LookupInput": {
        "properties": {
          "texts": {
            "items": {
              "type": "string",
              "maxLength": 10000
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Texts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "texts"
        ],
        "title": "LookupInput"
      },
      "LookupItem": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "translated": {
            "type": "string",
            "title": "Translated"
          },
          "matched": {
            "type": "boolean",
            "title": "Matched"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "translated",
          "matched",
          "groups"
        ],
        "title": "LookupItem"
      },
      "LookupOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LookupItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LookupOut"
      },
...
      "SessionOut": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "csrf": {
            "type": "string",
            "title": "Csrf"
          }
        },
        "type": "object",
        "required": [
          "user",
          "csrf"
        ],
        "title": "SessionOut"
      },
      "TranslationInput": {
        "properties": {
          "tag": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Tag"
          },
          "translation": {
            "type": "string",
            "maxLength": 2000,
            "title": "Translation",
            "default": ""
          },
          "groups": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 32,
            "title": "Groups"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tag"
        ],
        "title": "TranslationInput"
      },
      "TranslationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "translation": {
            "type": "string",
            "title": "Translation"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Groups"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tag",
          "translation",
          "groups",
          "revision",
          "source"
        ],
        "title": "TranslationOut"
      },
      "TranslationPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TranslationOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "TranslationPage"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "active"
        ],
        "title": "UserOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}